Add Keycloak auth integration and fix path aliases
